requirements.txt
djangorestframework==3.15.2
Django==5.0.6